requirements.txt
pynacl
pyyaml
git+https://github.com/tehzevo/protopost-python